+++ spec/_test_data/views/pages/mobile-sample/mobile-sample.js
@@ -1,0 +1,6 @@
+
+class MobileTestPage extends SipaBasicView {
+    onInit() {
+        console.warn("INIT MOBILE TEST PAGE");
+    }
+}