+++ spec/_test_data/views/pages/mobile-test/mobile-test.js
@@ -1,0 +1,6 @@
+
+class MobileSamplePage extends SipaBasicView {
+    onInit() {
+        console.warn("INIT MOBILE TEST PAGE");
+    }
+}