+++ spec/_test_data/views/pages/sample/sample.js
@@ -1,0 +1,6 @@
+
+class SamplePage extends SipaBasicView {
+    onInit() {
+        console.warn("INIT SAMPLE PAGE");
+    }
+}