+++ src/views/layouts/default/default.js
@@ -1,0 +1,9 @@
+class DefaultLayout extends SipaBasicView {
+    static onInit() {
+        // called when layout has been loaded, before (optional) fade animation
+    }
+
+    static onDestroy() {
+        // called when leaving layout, before next page will be loaded
+    }
+}