+++ src/views/pages/other/other.js
@@ -1,0 +1,13 @@
+class OtherPage extends SipaBasicView {
+    static onInit() {
+        // called when page has been loaded, before fade animation
+    }
+
+    static onShow() {
+        // called when page is visible to the user, after the init() and optional fade animation has completed
+    }
+
+    static onDestroy() {
+        // called when leaving page, before next page will be loaded
+    }
+}